Move `Identifier` to the callsite module
It makes more sense for it to live there.
