+++ tokio-trace-core/src/callsite.rs
@@ -1,6 +1,11 @@
 //! Callsites represent the source locations from which spans or events
 //! originate.
-use std::sync::Mutex;
+use std::{
+    fmt,
+    hash::{Hash, Hasher},
+    ptr,
+    sync::Mutex,
+};
 use {
     dispatcher::{self, Dispatch},
     subscriber::{Interest, Subscriber},
@@ -43,6 +48,12 @@ pub trait Callsite: Sync {
     fn metadata(&self) -> &Meta;
 }
 
+/// Uniquely identifies a [`Callsite`](::callsite::Callsite).
+///
+/// Two `Identifier`s are equal if they both refer to the same callsite.
+#[derive(Clone)]
+pub struct Identifier(&'static Callsite);
+
 /// Register a new `Callsite` with the global registry.
 ///
 /// This should be called once per callsite after the callsite has been constructed.
@@ -78,4 +89,47 @@ pub fn reset_registry() {
     let mut registry = REGISTRY.lock().unwrap();
     registry.callsites.clear();
     registry.dispatchers.clear();
+}
+
+// ===== impl Callsite =====
+
+impl Callsite + 'static {
+    /// Returns an `Identifier` unique to this `Callsite`.
+    // TODO: can this just be public API?
+    pub(crate) fn id(&'static self) -> Identifier {
+        Identifier::from_callsite(self)
+    }
+}
+
+// ===== impl Identifier =====
+
+impl Identifier {
+    /// Returns an `Identifier` unique to the provided `Callsite`.
+    // TODO: can this just be public API?
+    pub(crate) fn from_callsite(callsite: &'static Callsite) -> Self {
+        Identifier(callsite)
+    }
+}
+
+impl PartialEq for Identifier {
+    fn eq(&self, other: &Identifier) -> bool {
+        ptr::eq(self.0, other.0)
+    }
+}
+
+impl Eq for Identifier {}
+
+impl fmt::Debug for Identifier {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.pad("Identifier(...)")
+    }
+}
+
+impl Hash for Identifier {
+    fn hash<H>(&self, state: &mut H)
+    where
+        H: Hasher,
+    {
+        (self.0 as *const Callsite).hash(state)
+    }
 }
